vm: Add VM wrapper type and host_queries module
