+++ vm/tests/vm.rs
@@ -86,8 +86,7 @@ fn instantiate(vm: &VM, height: u64) -> (Session, ContractId) {
 
 #[test]
 fn hash() {
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
 
     let test_inputs = [
@@ -119,8 +118,7 @@ fn hash() {
 
 #[test]
 fn poseidon_hash() {
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
 
     let test_inputs = [
@@ -147,8 +145,7 @@ fn poseidon_hash() {
 
 #[test]
 fn schnorr_signature() {
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
 
     let sk = SchnorrSecretKey::random(&mut OsRng);
@@ -189,8 +186,7 @@ fn schnorr_signature() {
 
 #[test]
 fn bls_signature() {
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
 
     let message = b"some-message".to_vec();
@@ -222,8 +218,7 @@ fn bls_signature() {
 
 #[test]
 fn bls_multisig_signature() {
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
 
     let message = b"some-message".to_vec();
@@ -325,8 +320,7 @@ impl Circuit for PlonkTestCircuit {
 
 #[test]
 fn plonk_proof() {
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
 
     let pp = include_bytes!("./pp_test.bin");
@@ -404,8 +398,7 @@ impl<F: Groth16Field> ConstraintSynthesizer<F> for Groth16TestCircuit<F> {
 
 #[test]
 fn groth16_proof() {
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
 
     let test_circuit = Groth16TestCircuit {
@@ -497,8 +490,7 @@ fn groth16_proof() {
 fn chain_id() {
     const HEIGHT: u64 = 123;
 
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, HEIGHT);
 
     let chain_id: u8 = session
@@ -513,8 +505,7 @@ fn chain_id() {
 fn block_height() {
     const HEIGHT: u64 = 123;
 
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, HEIGHT);
 
     let height: u64 = session
@@ -535,8 +526,7 @@ fn get_owner() -> &'static BlsPublicKey {
 
 #[test]
 fn owner_raw() {
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
 
     let owner: [u8; 96] = session
@@ -549,8 +539,7 @@ fn owner_raw() {
 
 #[test]
 fn owner() {
-    let vm =
-        dusk_vm::new_ephemeral_vm().expect("Instantiating VM should succeed");
+    let vm = VM::ephemeral().expect("Instantiating VM should succeed");
     let (mut session, contract_id) = instantiate(&vm, 0);
 
     let owner: BlsPublicKey = session
